# openapi.with-responses.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/words/autocomplete": {
      "get": {
        "operationId": "WordsController_autocomplete",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "example": "resil",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 20,
              "default": 8,
              "example": 8,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/WordAutocompleteResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "400 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "401 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "500 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "tags": [
          "Words"
        ]
      }
    },
    "/api/v1/words/correct": {
      "post": {
        "operationId": "WordsController_correct",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectTextDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/TextCorrectionResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "400 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "401 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "500 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "tags": [
          "Words"
        ]
      }
    },
    "/api/v1/words/suggest": {
      "get": {
        "operationId": "WordsController_suggest",
        "parameters": [
          {
            "name": "word",
            "required": true,
            "in": "query",
            "schema": {
              "example": "apple",
              "type": "string"
            }
          },
          {
            "name": "targetLanguage",
            "required": false,
            "in": "query",
            "schema": {
              "default": "vi",
              "example": "vi",
              "type": "string"
            }
          },
          {
            "name": "imageLimit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 10,
              "default": 3,
              "example": 3,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/WordSuggestionResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "400 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "401 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "500 error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "tags": [
          "Words"
        ]
      }
    }
  },
  "info": {
    "title": "FLENVN API",
    "description": "API documentation for the FLENVN backend services Includes inferred response schemas for frontend integration. Successful responses are wrapped by ResponseInterceptor as { success, data, timestamp }.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT access token"
      }
    },
    "schemas": {
...
      "CorrectTextDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "example": "she go to school yesterday"
          },
          "language": {
            "type": "string",
            "example": "en-US",
            "default": "en-US"
          }
        },
        "required": [
          "text"
        ]
      },
      "ApiEnvelope": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {},
          "message": {
            "type": "string",
            "example": "Optional message"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-01T10:00:00.000Z"
          }
        },
        "required": [
          "success",
          "data",
          "timestamp"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-01T10:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/auth/login"
          },
          "method": {
            "type": "string",
            "example": "POST"
          },
          "message": {
            "type": "string",
            "example": "Invalid credentials"
          },
          "details": {}
        },
        "required": [
          "statusCode",
          "timestamp",
          "path",
          "method",
          "message"
        ]
      },
...
      "WordSuggestionResponse": {
        "type": "object",
        "properties": {
          "word": {
            "type": "string",
            "example": "serendipity"
          },
          "pronunciation": {
            "type": "string",
            "example": "/ˌserənˈdipəti/"
          },
          "partOfSpeech": {
            "type": "string",
            "example": "noun"
          },
          "definitions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string"
                },
                "partOfSpeech": {
                  "type": "string"
                }
              }
            }
          },
          "translation": {
            "type": "string",
            "example": "sự tình cờ may mắn"
          },
          "examples": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "example": "It was pure serendipity."
                },
                "translation": {
                  "type": "string",
                  "example": "Đó hoàn toàn là sự tình cờ may mắn."
                },
                "meaning": {
                  "type": "string"
                },
                "partOfSpeech": {
                  "type": "string"
                },
                "source": {
                  "type": "string",
                  "enum": [
                    "openai",
                    "dictionary"
                  ]
                }
              }
            }
          },
          "audio": {
            "type": "object",
            "properties": {
              "url": {
                "type": "string"
              },
              "source": {
                "type": "string",
                "enum": [
                  "polly",
                  "dictionary"
                ]
              }
            }
          },
          "images": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "url": {
                  "type": "string"
                },
                "source": {
                  "type": "string",
                  "enum": [
                    "pexels",
                    "unsplash",
                    "default"
                  ]
                },
                "photographer": {
                  "type": "string"
                },
                "photographerUrl": {
                  "type": "string"
                }
              }
            }
          },
          "sources": {
            "type": "object",
            "properties": {
              "dictionary": {
                "type": "string"
              },
              "translation": {
                "type": "string"
              },
              "audio": {
                "type": "string"
              },
              "examples": {
                "type": "string"
              },
              "images": {
                "type": "array",
                "items": {
                  "type": "string",
                  "enum": [
                    "pexels",
                    "unsplash",
                    "default"
                  ]
                }
              }
            }
          }
        }
      },
      "WordAutocompleteResponse": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string",
            "example": "resil"
          },
          "suggestions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "word": {
                  "type": "string",
                  "example": "resilient"
                },
                "score": {
                  "type": "number",
                  "example": 1234
                }
              }
            }
          },
          "source": {
            "type": "string",
            "enum": [
              "datamuse"
            ],
            "example": "datamuse"
          }
        }
      },
      "TextCorrectionResponse": {
        "type": "object",
        "properties": {
          "original": {
            "type": "string",
            "example": "she go to school yesterday"
          },
          "corrected": {
            "type": "string",
            "example": "She went to school yesterday."
          },
          "language": {
            "type": "string",
            "example": "en-US"
          },
          "suggestions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "offset": {
                  "type": "number",
                  "example": 4
                },
                "length": {
                  "type": "number",
                  "example": 2
                },
                "original": {
                  "type": "string",
                  "example": "go"
                },
                "replacements": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": [
                    "went"
                  ]
                },
                "message": {
                  "type": "string",
                  "example": "Possible grammar issue"
                },
                "ruleId": {
                  "type": "string",
                  "example": "PAST_TENSE"
                },
                "issueType": {
                  "type": "string",
                  "example": "grammar"
                }
              }
            }
          },
          "source": {
            "type": "string",
            "enum": [
              "languagetool"
            ],
            "example": "languagetool"
          }
        }
      }
    }
  }
}